+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu Sales
 ========================
 * dev-develop
 
+    * FEATURE     [CoreBundle]    Added option to not display currency in item-table.
     * BUGFIX      [CoreBundle]    Displaying first found tax class in item tables overlay.
     * FEATURE     [CoreBundle]    Display overlay (on click) also when item-table is disabled.
     * FEATURE     [CoreBundle]    Added recurring prices to item-table.